+++ src/App/Provider.php
@@ -1,4 +1,6 @@
-<?php # -*- coding: utf-8 -*-
+<?php
+
+# -*- coding: utf-8 -*-
 
 namespace Inpsyde\GoogleTagManager\App;