+++ src/Event/NoscriptTagRendererEvent.php
@@ -1,4 +1,8 @@
-<?php declare(strict_types=1); # -*- coding: utf-8 -*-
+<?php
+
+declare(strict_types=1);
+
+# -*- coding: utf-8 -*-
 
 namespace Inpsyde\GoogleTagManager\Event;
 
@@ -8,5 +12,5 @@ namespace Inpsyde\GoogleTagManager\Event;
 final class NoscriptTagRendererEvent
 {
 
-    const ACTION_RENDER = 'inpsyde-google-tag-manager.render-noscript';
+    public const ACTION_RENDER = 'inpsyde-google-tag-manager.render-noscript';
 }